Moved InvocationResult.java under results package
