+++ libevm/src/main/java/io/horizen/evm/results/InvocationResult.java
@@ -1,6 +1,7 @@
-package io.horizen.evm;
+package io.horizen.evm.results;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import io.horizen.evm.Address;
 
 import java.math.BigInteger;
 import java.util.Objects;
